+++ apps/next-client/src/components/medicine/FavoriteButton.tsx
@@ -5,6 +5,7 @@ import { useCheckFavorite, useAddFavorite } from '@/hooks/queries/useFavorite';
 import { ALERT_MESSAGES } from '@/constants/alertMessage';
 import { FavoriteButtonProps } from '@/dto/MedicineResultDto';
 import ErrorBoundary from '@/components/common/ErrorBoundary';
+import Cookies from 'js-cookie';
 
 const FavoriteButtonContent: React.FC<FavoriteButtonProps> = ({
   medicineId, itemName, entpName, etcOtcName, className, itemImage,
@@ -13,11 +14,16 @@ const FavoriteButtonContent: React.FC<FavoriteButtonProps> = ({
   const addFavoriteMutation = useAddFavorite();
 
   const handleAddFavorite = () => {
+    if (!Cookies.get('accessToken')) {
+      alert(ALERT_MESSAGES.ERROR.FAVORITES.LOGIN_REQUIRED);
+      return;
+    }
+
     if (isFavorite) {
       alert(ALERT_MESSAGES.SUCCESS.FAVORITE.FAVORITE_ALREADY_ADDED);
       return;
     }
-  
+
     addFavoriteMutation.mutate({
       medicineId,
       itemName,
@@ -26,23 +32,28 @@ const FavoriteButtonContent: React.FC<FavoriteButtonProps> = ({
       className: className ?? "",
       itemImage: itemImage ?? "",
     });
-  
+
     alert(ALERT_MESSAGES.SUCCESS.FAVORITE.FAVORITE_ADDED);
-  };  
+  };
 
   return (
     <button
       onClick={handleAddFavorite}
       className={`favorite_button ${isFavorite ? "active" : ""}`}
+      disabled={addFavoriteMutation.isPending}
     >
-      {isFavorite ? "⭐ 이미 추가됨" : "⭐ 즐겨찾기 추가"}
+      {addFavoriteMutation.isPending
+        ? "⭐ 처리 중..."
+        : isFavorite
+          ? "⭐ 이미 추가됨"
+          : "⭐ 즐겨찾기 추가"}
     </button>
   );
 };
 
 export const FavoriteButton: React.FC<FavoriteButtonProps> = (props) => (
   <ErrorBoundary>
-    <Suspense fallback={<button className="favorite_button">로딩 중...</button>}>
+    <Suspense fallback={<button className="favorite_button" disabled>⭐ 로딩 중...</button>}>
       <FavoriteButtonContent {...props} />
     </Suspense>
   </ErrorBoundary>
